Style and Behavioral Tweaks
First, we've updated the style of the discovery trail in presentation mode. The multiple borders were causing weird 1px off issues that resulted in a kludgy looking presenter mode. We've also updated the bookmarks interface so that the EditBookmark component highlights the title when the bookmark is active. We also detect clicks in the bookmark title to indicate it has been activated